Address PR feedback from May 1
